tools/ExamPDFRenamer/requirements.txt
PyMuPDF>=1.23.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
FreeSimpleGUI>=5.0.0
